警告回避。

+++ Classes/AKLibrary/AKInterface.cpp
@@ -391,7 +391,7 @@ AKLabel* AKInterface::addLabelMenu(const std::string menuString,
                                    bool withFrame)
 {
     // ラベルを作成する
-    AKLabel *label = AKLabel::createLabel(menuString, menuString.length(), 1, (withFrame ? kAKLabelFrameButton : kAKLabelFrameNone));
+    AKLabel *label = AKLabel::createLabel(menuString, (int)menuString.length(), 1, (withFrame ? kAKLabelFrameButton : kAKLabelFrameNone));
     
     // ラベルの位置を設定する
     label->setPosition(position);
